refactor: create useMultiPlay hook
